// client/src/api/generated-client_TEST.ts
export interface MemberLeftNotification extends BaseDto {
    clientId?: string;
    topic?: string;
}

export interface BatteryLevelDto extends BaseDto {
    batteryLevel?: number;
}

export interface DangerMovementDto extends BaseDto {
    command?: ClientCommandOfDistanceWarning;
}

export interface ClientCommandOfDistanceWarning {
    commandType?: ClientCommandType;
    payload?: DistanceWarning | undefined;
}

export enum ClientCommandType {
    Initialized = "initialized",
    BatteryInfo = "batteryInfo",
    DistanceWarning = "distanceWarning",
    NegativeWarning = "negativeWarning",
}

export interface DistanceWarning {
    warning?: string;
}

export interface InitializeEnginResponseDto extends BaseDto {
    command?: ClientCommandOfInitializeEngineResponse;
}

export interface ClientCommandOfInitializeEngineResponse {
    commandType?: ClientCommandType;
    payload?: InitializeEngineResponse | undefined;
}

export interface InitializeEngineResponse {
    initializeEngine?: boolean;
    errorMessage?: string;
}

export interface NegativeDistanceNotifierDto extends BaseDto {
    command?: ClientCommandOfNegativeDistanceWarning;
}

export interface ClientCommandOfNegativeDistanceWarning {
    commandType?: ClientCommandType;
    payload?: NegativeDistanceWarning | undefined;
}

export interface NegativeDistanceWarning {
    warning?: string;
}

export interface RobotMovementDto extends BaseDto {
    command?: CommandOfMovementCommand;
}

export interface CommandOfMovementCommand {
    commandType?: CommandType;
    payload?: MovementCommand | undefined;
}

export enum CommandType {
    Initialize = "initialize",
    Move = "move",
    Stop = "stop",
    Servo = "Servo",
}

export interface MovementCommand {
    directions?: Directions;
}

export interface Directions {
    activeMovements?: string[];
    lastCommand?: string;
}

export interface ServoDto extends BaseDto {
    command?: CommandOfServoCommand;
}

export interface CommandOfServoCommand {
    commandType?: CommandType;
    payload?: ServoCommand | undefined;
}

export interface ServoCommand {
    head?: number;
    neckt?: number;
    neckb?: number;
    leye?: number;
    reye?: number;
    lhand?: number;
    rhand?: number;
}

export interface ExampleServerResponse extends BaseDto {
    somethingTheServerSends?: string;
}

export interface ServerSendsErrorMessage extends BaseDto {
    message?: string;
}

export interface ServerSubscribedClientToRobotTopicDto extends BaseDto {
    subscribed?: boolean;
}

export interface ServerUnsubscribedClientFromRobotTopicDto extends BaseDto {
    unsubscribed?: boolean;
}

export interface Ping extends BaseDto {
}

export interface Pong extends BaseDto {
}

export interface ClientSendsTimerConfirmationDto extends BaseDto {
    status?: boolean;
    clientId?: string;
}

export interface ClientSubscribeToRobotTopicDto extends BaseDto {
    clientId?: string;
}

export interface ClientUnsubscribeFromRobotTopicDto extends BaseDto {
    clientId?: string;
}

export interface ExampleClientDto extends BaseDto {
    somethingTheClientSends?: string;
}

export interface EngineStateDto extends BaseDto {
    command?: CommandOfEngineManagement;
}

export interface CommandOfEngineManagement {
    commandType?: CommandType;
    payload?: EngineManagement | undefined;
}

export interface EngineManagement {
    engine?: boolean;
}

export interface ServerConfirmsDto extends BaseDto {
    success?: boolean;
}

export interface ServerSendsTimerDto extends BaseDto {
    status?: boolean;
    clientId?: string;
}

export interface ForceDisconnectedDto extends BaseDto {
    reason?: string;
}

/** Available eventType and string constants */
export enum StringConstants {
    MemberLeftNotification = "MemberLeftNotification",
    BatteryLevelDto = "BatteryLevelDto",
    DangerMovementDto = "DangerMovementDto",
    InitializeEnginResponseDto = "InitializeEnginResponseDto",
    NegativeDistanceNotifierDto = "NegativeDistanceNotifierDto",
    RobotMovementDto = "RobotMovementDto",
    ServoDto = "ServoDto",
    ExampleServerResponse = "ExampleServerResponse",
    ServerSendsErrorMessage = "ServerSendsErrorMessage",
    ServerSubscribedClientToRobotTopicDto = "ServerSubscribedClientToRobotTopicDto",
    ServerUnsubscribedClientFromRobotTopicDto = "ServerUnsubscribedClientFromRobotTopicDto",
    Ping = "Ping",
    Pong = "Pong",
    ClientSendsTimerConfirmationDto = "ClientSendsTimerConfirmationDto",
    ClientSubscribeToRobotTopicDto = "ClientSubscribeToRobotTopicDto",
    ClientUnsubscribeFromRobotTopicDto = "ClientUnsubscribeFromRobotTopicDto",
    ExampleClientDto = "ExampleClientDto",
    EngineStateDto = "EngineStateDto",
    ServerConfirmsDto = "ServerConfirmsDto",
    ServerSendsTimerDto = "ServerSendsTimerDto",
    ForceDisconnectedDto = "ForceDisconnectedDto",
}